+++ projects/ui-lib-custom/test/entry-points.spec.ts
@@ -8,6 +8,7 @@ import type * as UiLibColorPicker from 'ui-lib-custom/color-picker';
 import type * as UiLibCore from 'ui-lib-custom/core';
 import type * as UiLibDatePicker from 'ui-lib-custom/date-picker';
 import type * as UiLibEditor from 'ui-lib-custom/editor';
+import type * as UiLibFloatLabel from 'ui-lib-custom/float-label';
 import type * as UiLibIcon from 'ui-lib-custom/icon';
 import type * as UiLibInput from 'ui-lib-custom/input';
 import type * as UiLibLayout from 'ui-lib-custom/layout';
@@ -67,6 +68,11 @@ describe('Secondary Entry Points', (): void => {
   it('should import from editor', async (): Promise<void> => {
     const mod: typeof UiLibEditor = await import('ui-lib-custom/editor');
     expect(mod.EditorComponent).toBeDefined();
+  });
+
+  it('should import from float-label', async (): Promise<void> => {
+    const mod: typeof UiLibFloatLabel = await import('ui-lib-custom/float-label');
+    expect(mod.FloatLabelComponent).toBeDefined();
   });
 
   it('should import from icon', async (): Promise<void> => {